# swagger.yaml
openapi: 3.0.0
info:
  title: lightWaha - WhatsApp REST API
  description: |
    A lightweight WhatsApp integration service providing REST API endpoints to send and receive WhatsApp messages.

    Built with whatsapp-web.js and Express.js for direct WhatsApp Web automation.

    **Phase 2 Features (Active):**
    - ✅ Message Sending (POST /send)
    - ✅ Message Polling (GET /messages/new)
    - ✅ Message History (GET /messages)
    - ✅ Queue Statistics (GET /messages/stats)
    - ✅ WhatsApp Authentication (QR code)

    **Future (Phase 3/4):**
    - 🔜 Webhook-based message delivery (real-time)
    - 🔜 Message acknowledgment tracking
  version: 1.0.0
  contact:
    name: lightWaha Support
  license:
    name: MIT
servers:
  - url: http://localhost:4000
    description: Local development server
  - url: http://192.168.0.116:4000
    description: Production server example
...
paths:
...
  /qr:
    get:
      summary: Get QR code for authentication (JSON)
      description: |
        Returns the current QR code as a string. Scan this with WhatsApp Linked Devices to authenticate.

        **Note:** Only available when not yet authenticated. Returns error once authenticated.
      tags:
        - Authentication
      responses:
        '200':
          description: QR code available for scanning
          content:
            application/json:
              schema:
                type: object
                properties:
                  qr:
                    type: string
                    description: QR code data string
                  message:
                    type: string
                    example: "QR code generated. Check server terminal for visual display."
        '400':
          description: No QR code available (already authenticated)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No QR code available. Already authenticated?"
                  hint:
                    type: string
                    example: "Check the server terminal for the QR code display"
  /qr.html:
    get:
      summary: Get QR code authentication page (HTML)
      description: |
        Returns an interactive HTML page with:
        - Visual QR code display
        - Step-by-step authentication instructions
        - Link to check connection status

        **Perfect for:** Web-based deployment where you can open a browser to scan the QR code
      tags:
        - Authentication
      responses:
        '200':
          description: HTML page with QR code and instructions
          content:
            text/html:
              schema:
                type: string
              example: |
                <!DOCTYPE html>
                <html>
                <head><title>lightWaha - QR Code</title></head>
                <body>
                  <div class="container">
                    <h1>📱 lightWaha</h1>
                    <img src="...qr-code-image...">
                    <ol>
                      <li>Open WhatsApp on your mobile device</li>
                      <li>Go to Settings</li>
                      <li>Select Linked Devices</li>
                      <li>Tap Link a device</li>
                      <li>Point your camera at the QR code above</li>
                    </ol>
                  </div>
                </body>
                </html>
        '400':
          description: No QR code available (already authenticated)
          content:
            text/html:
              schema:
                type: string
  /messages/new:
    get:
      summary: Get new messages (polling endpoint)
      description: |
        **Polling-based message retrieval (Phase 2 - Option 4)**

        Returns all messages received since the given timestamp.
        Use this to implement client-side polling for new messages.

        **Workflow:**
        1. Get messages since timestamp 0: `GET /messages/new?since=0`
        2. Process messages
        3. Store the `cursor` value
        4. Next poll: `GET /messages/new?since={cursor}`

        **Cursor-based pagination:**
        - Returns `cursor` = latest message timestamp
        - Use as `since` parameter for next request
        - Guarantees no message gaps or duplicates
      tags:
        - Messaging
      parameters:
        - name: since
          in: query
          schema:
            type: integer
            default: 0
          description: |
            Unix timestamp in milliseconds.
            Returns messages with timestamp > since.
            Use 0 to get all messages.
          example: 1646000000000
      responses:
        '200':
          description: Successful message retrieval
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          example: "true_351910270614@c.us_3EB08BE77018B4BF937709"
                        from:
                          type: string
                          example: "351910270614@c.us"
                        fromName:
                          type: string
                          example: "User Name"
                        body:
                          type: string
                          example: "Hello! How are you?"
                        timestamp:
                          type: number
                          example: 1646000000000
                        isFromMe:
                          type: boolean
                          example: false
                  cursor:
                    type: number
                    description: Latest message timestamp (use as 'since' for next poll)
                    example: 1646000000100
                  count:
                    type: integer
                    example: 2
        '400':
          description: Invalid 'since' parameter
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /messages:
    get:
      summary: Get message history with pagination
      description: |
        Retrieve stored messages with pagination support.

        **Parameters:**
        - `limit`: Messages to return per request (max 500)
        - `offset`: Number of messages to skip from start

        **Example:**
        ```
        GET /messages?limit=20&offset=0   # First 20 messages
        GET /messages?limit=20&offset=20  # Next 20 messages
        ```

        **Note:** This returns all stored messages (both incoming and outgoing).
      tags:
        - Messaging
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 500
          description: Maximum number of messages to return
          example: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Number of messages to skip from the beginning
          example: 0
      responses:
        '200':
          description: Message history with pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 150
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        from:
                          type: string
                        fromName:
                          type: string
                        body:
                          type: string
                        timestamp:
                          type: number
                        isFromMe:
                          type: boolean
                  limit:
                    type: integer
                    example: 50
                  offset:
                    type: integer
                    example: 0
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
  /send:
    post:
      summary: Send a WhatsApp message
      description: |
        Send a text message to a WhatsApp contact via REST API.

        **Requirements:**
        - WhatsApp must be authenticated (check `/status`)
        - Recipient must be a valid phone number (with or without +)
        - Message text must not be empty

        **Phone Number Formats Accepted:**
        - `351910270614` (without country code)
        - `+351910270614` (with country code)
        - Both formats are normalized internally
      tags:
        - Messaging
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - to
                - text
              properties:
                to:
                  type: string
                  description: Recipient phone number (with or without + prefix)
                  example: "351910270614"
                text:
                  type: string
                  description: Message content to send
                  example: "Hello! This is a test message"
      responses:
        '200':
          description: Message sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: object
                    description: WhatsApp message metadata
                    properties:
                      id:
                        type: string
                        example: "3EB08BE77018B4BF937709"
                      fromMe:
                        type: boolean
                        example: true
                      remote:
                        type: string
                        example: "212244549910628@lid"
                      _serialized:
                        type: string
                        example: "true_212244549910628@lid_3EB08BE77018B4BF937709"
                  to:
                    type: string
                    example: "+351910270614"
                  text:
                    type: string
                    example: "Hello! This is a test message"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2026-03-08T21:19:19.138Z"
        '400':
          description: Invalid request or WhatsApp not authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Missing required fields: to, text"
        '500':
          description: Failed to send message
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to send message"
                  details:
                    type: string
                    example: "Contact not found or chat error"
  /logout:
    post:
      summary: Logout from WhatsApp
      description: |
        Disconnect the WhatsApp session and clear authentication.

        **Effect:** Clears stored session credentials. Next request will require re-scanning QR code.
      tags:
        - Session Management
      responses:
        '200':
          description: Successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Logged out from WhatsApp"
        '500':
          description: Failed to logout
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to logout"
                  details:
                    type: string
  /destroy:
    post:
      summary: Destroy WhatsApp client
      description: |
        Completely destroy the WhatsApp client instance and clean up resources.

        **Use Case:** Server shutdown or complete reset. More aggressive than `/logout`.
      tags:
        - Session Management
      responses:
        '200':
          description: Client successfully destroyed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "WhatsApp client destroyed"
        '500':
          description: Failed to destroy client
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to destroy client"
                  details:
                    type: string
...